% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tertiles.R
\name{assign_tertiles}
\alias{assign_tertiles}
\title{assign_tertiles
Divides continuous gene expression levels into 3 tertiles and label them "_high", "_medium" and "_low" concatenating with the input id text.}
\usage{
assign_tertiles(numeric_vector1, id1)
}
\arguments{
\item{numeric_vector1}{any continuous numeric vector}

\item{id1}{A character of length 1. Used to describe the tertiles}
}
\value{
character vector of the same length with input numeric vector. The values are replaced by 3 levels of tertiles.
}
\description{
assign_tertiles
Divides continuous gene expression levels into 3 tertiles and label them "_high", "_medium" and "_low" concatenating with the input id text.
}
\examples{
tertiles1<-assign_tertiles(numeric_vector1, "inflamed")
}
\references{
\url{http://github.com/tolgaturan-github/IBRIDGE}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
