% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify_cells.R
\name{Classify_cells}
\alias{Classify_cells}
\title{Classify single-cells or in vitro cell lines
Classifies the input single-cell or in vitro cell line data into "inflamed", "cold" or "unassigned" bins}
\usage{
Classify_cells(
  seu1,
  iBRIDGE_features,
  norm_method = "SCTransform",
  n_cores = 1
)
}
\arguments{
\item{seu1}{Output of UMI count matrix normalized by Seurat workflows}

\item{iBRIDGE_features}{Output of IBRIDGE_overlaps function.}

\item{norm_method}{"SCTransform" or "NormalizeData". Defaults to "SCTransform"}

\item{n_cores}{Integer of length 1. Specifies the number of parallel cores running the function.
Defaults to 1}
}
\value{
A seurat_object with metadata updated with iBRIDGE classification
}
\description{
Classify single-cells or in vitro cell lines
Classifies the input single-cell or in vitro cell line data into "inflamed", "cold" or "unassigned" bins
}
\examples{
seu1<-Classify_cells(seu1, iBRIDGE_feautures, "SCTransform", 100)
}
\references{
\url{http://github.com/tolgaturan-github/IBRIDGE}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
