% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBRIDGE_overlaps.R
\name{IBRIDGE_overlaps}
\alias{IBRIDGE_overlaps}
\title{IBRIDGE_Overlaps
Integrates bulk with single-cell (or cell line) gene expression data and identifies features that can classfy cells into "inflamed" and "cold" bins.}
\usage{
IBRIDGE_overlaps(
  seurat_object,
  type = "ens",
  cohort = c("ACC", "BLCA", "CESC", "CHOL", "COAD", "ESCA", "GBM", "HNSC", "KICH",
    "KIRC", "KIRP", "LGG", "LIHC", "LUAD", "LUSC", "MESO", "OV", "PAAD", "PRAD", "READ",
    "SARC", "SKCM", "STAD", "THCA", "UCEC", "UCS"),
  norm_method = "SCTransform"
)
}
\arguments{
\item{seurat_object}{Output of UMI count matrix normalized by Seurat workflows.}

\item{type}{type of gene identifier the dataset has. "ens" for ensembl IDs and "gene" for HUGO gene exymbols. Defaults to "ens"}

\item{cohort}{one of 28 solid tumor cohorts from TCGA database. The cancer cohort from which the single-cell gene expression dataset is generated.}

\item{norm_method}{"SCTransform" or "NormalizeData". Defaults to "SCTransform"}
}
\value{
list of 2 elements. One for "inflamed" and one for "cold" features.
}
\description{
IBRIDGE_Overlaps
Integrates bulk with single-cell (or cell line) gene expression data and identifies features that can classfy cells into "inflamed" and "cold" bins.
}
\examples{
IBRIDGE_features<-IBRIDGE_overlaps(seu1, "LUAD", "SCTransform")
}
\references{
\url{http://github.com/tolgaturan-github/IBRIDGE}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
