% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_enrichment.R
\name{sc_geneset_enrich}
\alias{sc_geneset_enrich}
\title{Geneset Enrichment (GSE) on scRNAseq Data
Calculates single cell GSE scores using "AUCell" method}
\usage{
sc_geneset_enrich(
  seurat_object,
  geneset_list,
  norm_method = "SCTransform",
  n_cores = 1
)
}
\arguments{
\item{seurat_object}{Output of UMI count matrix normalized by Seurat workflows}

\item{geneset_list}{a List of gene/Transcript features.
Preferably all or at least 1 feature in each geneset must be present in the rownames of the expression matrix.}

\item{norm_method}{"SCTransform" or "NormalizeData". Defaults to "SCTransform"}

\item{n_cores}{Integer of length 1. Specifies the number of parallel cores running the function.
Defaults to 1}
}
\value{
data.frame holding geneset enrichment values for each geneset (columns) and for each cell (rows).
}
\description{
Geneset Enrichment (GSE) on scRNAseq Data
Calculates single cell GSE scores using "AUCell" method
}
\examples{
enrichment_scores1<-sc_geneset_enrich(seu1, geneset_list1, "SCTransform", 100)
}
\references{
\url{http://github.com/tolgaturan-github/IBRIDGE}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
