% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCTransform_norm.R
\name{SCTransform_normalization}
\alias{SCTransform_normalization}
\title{SCTransform based scRNAseq Workflow
This function accepts UMI count matrix and patient metadata as arguments
and replaces NormalizeData(), ScaleData(), and FindVariableFeatures() functions from the Seurat package
The function also performs UMAP dimensio reduction and clustering}
\usage{
SCTransform_normalization(countmatrix, metadata)
}
\arguments{
\item{countmatrix}{Numeric matrix of UMI counts genes as rows and cell_ids as columns from malignant subset of the scRNAseq data
This matrix is usually a subset of a scRNAseq dataset profiling the whole TIME.}

\item{metadata}{Character vector (1-dimension) of length matching the cell count for the input countmatrix
Preferably all or at least 1 feature in each geneset must be present in the rownames of the expression matrix.}
}
\value{
A seurat object with SCTransform normalized counts and PErson residuals as well as patient metadata and dimension reduction slots
}
\description{
SCTransform based scRNAseq Workflow
This function accepts UMI count matrix and patient metadata as arguments
and replaces NormalizeData(), ScaleData(), and FindVariableFeatures() functions from the Seurat package
The function also performs UMAP dimensio reduction and clustering
}
\examples{
seurat_object1<-SCTransform_normalization(countmatrix1, patient_metadata1)
}
\references{
\url{http://github.com/tolgaturan-github/IBRIDGE}
}
\author{
Tolga Turan, \email{tolga.turan@abbvie.com}
}
